requirements.txt
asgiref==3.10.0
cffi==2.0.0
cryptography==46.0.2
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-rest-knox==4.2.0
django-rest-passwordreset==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pycparser==2.23
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
